% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chyrons.R
\docType{package}
\name{Chyrons}
\alias{Chyrons}
\alias{Chyrons-package}
\title{Chyrons}
\format{
A dataframe with 21306 elements
\describe{
\item{date_time_utc}{Date and time in UTC in the format M/DD/YYYY H:MM}
\item{channel}{News channel (of the chyron)}
\item{duration}{Time (in seconds) the chyron showed on screen}
\item{text}{Text of the chyron}
}
}
\source{
\url{https://archive.org/services/third-eye.php}
}
\usage{
Chyrons
}
\description{
This package contains information regarding the variables
for Chyrons

A data set containing information regarding chyrons
from the time period 9/23/20 to 10/4/20.
}
\keyword{datasets}
